requirements.txt
line-bot-sdk==3.5.0
fastapi
uvicorn[standard]
google.generativeai
yfinance
pydantic
tiktoken
Pillow
python-dotenv
pytesseract 
requests
langchain
langchain-google-genai
langchain-community
git+https://github.com/ozgur/python-firebase